refactor(cli): keep Registry v3 behind existing commands
