+++ internal/cloud/client.go
@@ -34,41 +34,6 @@ type PackageVersionRecord struct {
 	CreatedAt string `json:"created_at"`
 }
 
-type PackageRecord struct {
-	Scope         string                 `json:"scope"`
-	Name          string                 `json:"name"`
-	Ref           string                 `json:"ref"`
-	DisplayName   *string                `json:"display_name,omitempty"`
-	Description   *string                `json:"description,omitempty"`
-	Visibility    string                 `json:"visibility"`
-	Tags          []string               `json:"tags"`
-	LatestVersion *PackageVersionRecord  `json:"latest_version,omitempty"`
-	Versions      []PackageVersionRecord `json:"versions,omitempty"`
-	CreatedAt     string                 `json:"created_at"`
-	UpdatedAt     string                 `json:"updated_at"`
-	CanManage     bool                   `json:"can_manage"`
-}
-
-type RegistryVisibilityBlocker struct {
-	Code    string  `json:"code"`
-	Message string  `json:"message"`
-	Ref     *string `json:"ref,omitempty"`
-}
-
-type RegistryVisibilityPreflight struct {
-	ArtifactKind      string                      `json:"artifact_kind"`
-	Scope             string                      `json:"scope"`
-	Name              string                      `json:"name"`
-	CurrentVisibility string                      `json:"current_visibility"`
-	TargetVisibility  string                      `json:"target_visibility"`
-	IdentityRevision  int                         `json:"identity_revision"`
-	VersionCount      int                         `json:"version_count"`
-	VersionSetDigest  string                      `json:"version_set_digest"`
-	ConfirmationToken string                      `json:"confirmation_token"`
-	Warning           *string                     `json:"warning,omitempty"`
-	Blockers          []RegistryVisibilityBlocker `json:"blockers"`
-}
-
 type Capabilities struct {
 	DeploymentRevisionHistory  bool `json:"deployment_revision_history"`
 	DeploymentRevisionMessages bool `json:"deployment_revision_messages"`
@@ -94,14 +59,6 @@ type SkillRecord struct {
 	SourceRef   string         `json:"source_ref"`
 	Visibility  string         `json:"visibility"`
 	CanManage   bool           `json:"can_manage"`
-}
-
-type packageListResponse struct {
-	Packages []PackageRecord `json:"packages"`
-}
-
-type skillListResponse struct {
-	Skills []SkillRecord `json:"skills"`
 }
 
 type SessionRecord struct {
@@ -303,10 +260,8 @@ func ControlClientForContext(contextOverride string) (*Client, error) {
 	return client, nil
 }
 
-// RegistryReadClientForContext returns an authenticated Registry client when
-// credentials are configured and otherwise returns an anonymous client for
-// public Registry reads. An explicitly selected organization always requires
-// authentication.
+// RegistryReadClientForContext permits anonymous reads only when no
+// organization context was explicitly selected.
 func RegistryReadClientForContext(contextOverride string) (*Client, error) {
 	cfg, err := config.LoadConfig()
 	if err != nil {
@@ -368,41 +323,6 @@ func (c *Client) PublishPackageWithVisibility(
 		return nil, readError(resp)
 	}
 	var record PackageVersionRecord
-	if err := json.NewDecoder(resp.Body).Decode(&record); err != nil {
-		return nil, err
-	}
-	return &record, nil
-}
-
-func (c *Client) ListPackages() ([]PackageRecord, error) {
-	resp, err := c.do("GET", "/api/packages", nil)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var response packageListResponse
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		return nil, err
-	}
-	return response.Packages, nil
-}
-
-func (c *Client) GetPackage(scope, name string) (*PackageRecord, error) {
-	path := "/api/packages/" +
-		url.PathEscape(strings.TrimSpace(scope)) + "/" +
-		url.PathEscape(strings.TrimSpace(name))
-	resp, err := c.do("GET", path, nil)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var record PackageRecord
 	if err := json.NewDecoder(resp.Body).Decode(&record); err != nil {
 		return nil, err
 	}
@@ -499,22 +419,6 @@ func (c *Client) PublishSkillVersionWithVisibility(
 	return &record, nil
 }
 
-func (c *Client) ListSkills() ([]SkillRecord, error) {
-	resp, err := c.do("GET", "/api/skills", nil)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var response skillListResponse
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		return nil, err
-	}
-	return response.Skills, nil
-}
-
 func (c *Client) GetSkill(scope, name string) (*SkillRecord, error) {
 	path := "/api/skills/" + url.PathEscape(strings.TrimSpace(scope)) + "/" + url.PathEscape(strings.TrimSpace(name))
 	return c.getSkill(path)
@@ -592,115 +496,6 @@ func (c *Client) RegistryCapabilities() (*Capabilities, error) {
 		return nil, err
 	}
 	return &capabilities, nil
-}
-
-func (c *Client) PreflightPackageVisibility(
-	scope string,
-	name string,
-	visibility string,
-) (*RegistryVisibilityPreflight, error) {
-	return c.preflightRegistryVisibility("packages", scope, name, visibility)
-}
-
-func (c *Client) PreflightSkillVisibility(
-	scope string,
-	name string,
-	visibility string,
-) (*RegistryVisibilityPreflight, error) {
-	return c.preflightRegistryVisibility("skills", scope, name, visibility)
-}
-
-func (c *Client) preflightRegistryVisibility(
-	kind string,
-	scope string,
-	name string,
-	visibility string,
-) (*RegistryVisibilityPreflight, error) {
-	body, err := json.Marshal(map[string]string{
-		"visibility": strings.TrimSpace(visibility),
-	})
-	if err != nil {
-		return nil, err
-	}
-	path := "/api/" + kind + "/" +
-		url.PathEscape(strings.TrimSpace(scope)) + "/" +
-		url.PathEscape(strings.TrimSpace(name)) + "/visibility/preflight"
-	resp, err := c.do("POST", path, body)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var preflight RegistryVisibilityPreflight
-	if err := json.NewDecoder(resp.Body).Decode(&preflight); err != nil {
-		return nil, err
-	}
-	return &preflight, nil
-}
-
-func (c *Client) ChangePackageVisibility(
-	scope string,
-	name string,
-	visibility string,
-	confirmationToken string,
-) (*PackageRecord, error) {
-	body, err := registryVisibilityApplyBody(visibility, confirmationToken)
-	if err != nil {
-		return nil, err
-	}
-	path := "/api/packages/" +
-		url.PathEscape(strings.TrimSpace(scope)) + "/" +
-		url.PathEscape(strings.TrimSpace(name)) + "/visibility"
-	resp, err := c.do("PUT", path, body)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var record PackageRecord
-	if err := json.NewDecoder(resp.Body).Decode(&record); err != nil {
-		return nil, err
-	}
-	return &record, nil
-}
-
-func (c *Client) ChangeSkillVisibility(
-	scope string,
-	name string,
-	visibility string,
-	confirmationToken string,
-) (*SkillRecord, error) {
-	body, err := registryVisibilityApplyBody(visibility, confirmationToken)
-	if err != nil {
-		return nil, err
-	}
-	path := "/api/skills/" +
-		url.PathEscape(strings.TrimSpace(scope)) + "/" +
-		url.PathEscape(strings.TrimSpace(name)) + "/visibility"
-	resp, err := c.do("PUT", path, body)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, readError(resp)
-	}
-	var record SkillRecord
-	if err := json.NewDecoder(resp.Body).Decode(&record); err != nil {
-		return nil, err
-	}
-	return &record, nil
-}
-
-func registryVisibilityApplyBody(visibility, confirmationToken string) ([]byte, error) {
-	return json.Marshal(map[string]string{
-		"visibility":         strings.TrimSpace(visibility),
-		"confirmation_token": strings.TrimSpace(confirmationToken),
-	})
 }
 
 func (c *Client) CreateSession(opts SessionCreateOptions) (*SessionRecord, error) {
