+++ internal/cloud/client_test.go
@@ -165,30 +165,11 @@ func TestControlClientUsesStableContextWithoutLookup(t *testing.T) {
 	}
 }
 
-func TestRegistryReadClientAllowsAnonymousReadsButNotExplicitOrgContext(t *testing.T) {
-	t.Setenv(config.ConfigPathEnv, filepath.Join(t.TempDir(), "missing.yaml"))
-	t.Setenv(config.APIEndpointEnv, "https://api.example.com")
-	t.Setenv(config.AuthTokenEnv, "")
-	t.Setenv(config.ContextEnv, "@stored-team")
-
-	client, err := RegistryReadClientForContext("")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if client.Token != "" || client.OrgID != "" {
-		t.Fatalf("anonymous Registry client = %+v", client)
-	}
-	if _, err := RegistryReadClientForContext("@explicit-team"); err == nil ||
-		!strings.Contains(err.Error(), "requires login") {
-		t.Fatalf("explicit anonymous org context error = %v", err)
-	}
-}
-
-func TestClientRegistryPrivacyCommandsUseExplicitWireContracts(t *testing.T) {
+func TestClientRegistryPrivacyPublishingUsesExplicitWireContracts(t *testing.T) {
 	requests := map[string]map[string]any{}
 	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		key := r.Method + " " + r.URL.Path
-		if r.Method == http.MethodPost || r.Method == http.MethodPut {
+		if r.Method == http.MethodPost {
 			var body map[string]any
 			if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
 				t.Fatal(err)
@@ -202,19 +183,6 @@ func TestClientRegistryPrivacyCommandsUseExplicitWireContracts(t *testing.T) {
 			_ = json.NewEncoder(w).Encode(map[string]any{
 				"scope": "alice", "name": "demo", "version": "1.0.0",
 				"ref": "@alice/demo:1.0.0", "digest": "sha256:package",
-			})
-		case "POST /api/packages/alice/demo/visibility/preflight":
-			_ = json.NewEncoder(w).Encode(map[string]any{
-				"artifact_kind": "package", "scope": "alice", "name": "demo",
-				"current_visibility": "private", "target_visibility": "public",
-				"identity_revision": 1, "version_count": 2,
-				"version_set_digest": "sha256:versions",
-				"confirmation_token": "sha256:confirmation", "blockers": []any{},
-			})
-		case "PUT /api/packages/alice/demo/visibility":
-			_ = json.NewEncoder(w).Encode(map[string]any{
-				"scope": "alice", "name": "demo", "ref": "@alice/demo",
-				"visibility": "public",
 			})
 		case "POST /api/skills":
 			_ = json.NewEncoder(w).Encode(map[string]any{
@@ -238,15 +206,6 @@ func TestClientRegistryPrivacyCommandsUseExplicitWireContracts(t *testing.T) {
 	); err != nil {
 		t.Fatal(err)
 	}
-	preflight, err := client.PreflightPackageVisibility("alice", "demo", "public")
-	if err != nil || preflight.VersionCount != 2 {
-		t.Fatalf("preflight = %+v, %v", preflight, err)
-	}
-	if _, err := client.ChangePackageVisibility(
-		"alice", "demo", "public", preflight.ConfirmationToken,
-	); err != nil {
-		t.Fatal(err)
-	}
 	files := map[string]SkillFile{"SKILL.md": {Mode: "0644", Data: []byte("skill")}}
 	if _, err := client.PublishSkillVersionWithVisibility(
 		"alice", "lint", "1.0.0", files, "public",
@@ -255,10 +214,6 @@ func TestClientRegistryPrivacyCommandsUseExplicitWireContracts(t *testing.T) {
 	}
 	if requests["POST /api/packages"]["visibility"] != "public" {
 		t.Fatalf("package publish body = %#v", requests["POST /api/packages"])
-	}
-	if requests["PUT /api/packages/alice/demo/visibility"]["confirmation_token"] !=
-		"sha256:confirmation" {
-		t.Fatalf("visibility body = %#v", requests["PUT /api/packages/alice/demo/visibility"])
 	}
 	if requests["POST /api/skills"]["visibility"] != "public" {
 		t.Fatalf("skill publish body = %#v", requests["POST /api/skills"])
